Implementing validations, exceptions and service modules

+++ backend/src/barbers/barbers.service.ts
@@ -1,6 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import { isValidObjectId, Model } from 'mongoose';
 import { Barber } from './schemas/barber.schema';
 import { CreateBarberDto } from './dto/create-barber.dto';
 import { UpdateBarberDto } from './dto/update-barber.dto';
@@ -19,20 +23,52 @@ export class BarbersService {
   }
 
   async findByName(name: string): Promise<Barber> {
-    return this.barberModel.findOne({ name }).exec();
+    const barber = await this.barberModel.findOne({ name }).exec();
+    if (!barber) {
+      throw new NotFoundException(`Barber with name ${name} not found`);
+    }
+
+    return barber;
   }
 
   async findOne(id: string): Promise<Barber> {
-    return this.barberModel.findById(id).exec();
+    this.validateObjectId(id);
+
+    const barber = await this.barberModel.findById(id).exec();
+    if (!barber) {
+      throw new NotFoundException(`Barber with Id ${id} not found`);
+    }
+
+    return barber;
   }
 
   async update(id: string, updateBarberDto: UpdateBarberDto): Promise<Barber> {
+    this.validateObjectId(id);
+
+    const existingBarber = await this.barberModel.findById(id).exec();
+    if (!existingBarber) {
+      throw new NotFoundException(`Barber with Id ${id} not found`);
+    }
+
     return this.barberModel
       .findByIdAndUpdate(id, updateBarberDto, { new: true })
       .exec();
   }
 
   async remove(id: string): Promise<Barber> {
+    this.validateObjectId(id);
+
+    const existingBarber = await this.barberModel.findById(id).exec();
+    if (!existingBarber) {
+      throw new NotFoundException(`Barber with Id ${id} not found`);
+    }
+
     return this.barberModel.findByIdAndDelete(id).exec();
+  }
+
+  private validateObjectId(id: string): void {
+    if (!isValidObjectId(id)) {
+      throw new BadRequestException(`Invalid Id format: ${id}`);
+    }
   }
 }
